Commit-8-ServiceImplementationAndPartialServiceTests

+++ src/main/java/DAO/MessageDAO.java
@@ -10,32 +10,27 @@ import java.util.List;
 public class MessageDAO {
     public static boolean test = true;
 
-    //User Story 3 method (Remove Service Logic)
+    //User Story 3 DAO Method (Remove Service Logic)
     public Message createMessage(int posted_by, String message_text, long time_posted_epoch){
         Connection conn = ConnectionUtil.getConnection();
-        //AccountDAO method access
-        AccountDAO accDao = new AccountDAO();
-        if(message_text.length()<=255 && accDao.checkAccountExistsByID(posted_by)){
-            try{
-                String sql = "INSERT INTO message (posted_by,message_text,time_posted_epoch) VALUES (?,?,?)";
-                PreparedStatement ps = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
-                ps.setInt(1, posted_by);
-                ps.setString(2, message_text);
-                ps.setLong(3, time_posted_epoch);
-                ps.executeUpdate();
-                ResultSet pkeys = ps.getGeneratedKeys();
-                pkeys.next();
-                int generated_account_id = (int) pkeys.getInt(1);
-                return new Message(generated_account_id, posted_by, message_text, time_posted_epoch);
-            }catch(SQLException e){
-                System.out.println(e.getMessage());
-                // System.out.println("lala");
-            }
+        try{
+            String sql = "INSERT INTO message (posted_by,message_text,time_posted_epoch) VALUES (?,?,?)";
+            PreparedStatement ps = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
+            ps.setInt(1, posted_by);
+            ps.setString(2, message_text);
+            ps.setLong(3, time_posted_epoch);
+            ps.executeUpdate();
+            ResultSet pkeys = ps.getGeneratedKeys();
+            pkeys.next();
+            int generated_account_id = (int) pkeys.getInt(1);
+            return new Message(generated_account_id, posted_by, message_text, time_posted_epoch);
+        }catch(SQLException e){
+            System.out.println(e.getMessage());
         }
         return null;
     }
 
-    //User Story 5 method, User Story 6 Method, User Story 7 Method
+    //User Story 5 DAO Method, User Story 6 DAO Method, User Story 7 DAO Method
     public Message getMessageByID(int id){
         Connection conn = ConnectionUtil.getConnection();
         try{
@@ -55,20 +50,25 @@ public class MessageDAO {
         }
         return null;    }
 
-    //User Story 6 method
-    public void deleteMessageByID(int id){
+    //User Story 6 DAO Method
+    public Message deleteMessageByID(int id){
         Connection conn = ConnectionUtil.getConnection();
-        try{
-            String sql = "DELETE FROM message WHERE message_id = ?";
-            PreparedStatement ps = conn.prepareStatement(sql);
-            ps.setInt(1, id);
-            ps.executeUpdate();
-        }catch(SQLException e){
-            System.out.println(e.getMessage());
+        if (getMessageByID(id)!=null){
+            try{
+                Message del_msg = getMessageByID(id);
+                String sql = "DELETE FROM message WHERE message_id = ?";
+                PreparedStatement ps = conn.prepareStatement(sql);
+                ps.setInt(1, id);
+                ps.executeUpdate();
+                return del_msg;
+            }catch(SQLException e){
+                System.out.println(e.getMessage());
+            }
         }
+        return null;
     }
 
-    //User Story 7 method
+    //User Story 7 DAO Method
     public Message updateMessageByID(int id, String text){
         Connection conn = ConnectionUtil.getConnection();
         try{
@@ -85,7 +85,7 @@ public class MessageDAO {
         return null;
     }
 
-    //User Story 8 method
+    //User Story 8 DAO Method
     public List<Message> getMessagesByUserID(int id){
         Connection conn = ConnectionUtil.getConnection();
         List<Message> messages = new ArrayList<>();
@@ -108,7 +108,7 @@ public class MessageDAO {
         return messages;
     }
 
-    //User Story 4 method
+    //User Story 4 DAO Method
     public List<Message> getAllMessages(){
         Connection conn = ConnectionUtil.getConnection();
         List<Message> messages = new ArrayList<>();
